project: support config using apkg.toml
Improve Project handling and add configuration support.

New project config file is loaded by default:

    distro/config/apkg.toml

It's written in toml format (like ini but better) using toml python
module which is a new apkg requirement.

toml looks good on paper but if it proves inferior to YAML in practice,
it will be replaced.