log: improve printing of command outputs
Don't redirect stdout/stderr (direct run() mode) by default
for important commands for building (source) packages.

This means normal output is shown in terminal when packaging commands
are run as if user ran them manually.

Output can be supressed with --brief or --quiet.

Note that direct mode is never used for make_archive_script because
its output must be redirected in order to be parsed.