build: use isolated build by default
It's dangerous for new users to use direct build by default as it
modifies host system and requires build deps to be installed
system-wide. This is useful especially in VMs and containers (CI).

Isolated build is safer so make it a default with -H/--host-build
option to enable direct host build replacing --isolated option.

Adjust arch pkgstyle to work with new default CI invocation even though
build deps installation and direct host build aren't supported on arch:

    apkg build -Hi

apkg will print warnings for -H/--host-build (not supported on arch in
general) and -i/--install-deps (not needed on arch) but it's going to
work regardless.

    apkg build-dep

on arch will result in DistroNotSupported error (exit code 44) as it
should.

Fixes: https://gitlab.nic.cz/packaging/apkg/-/issues/27